requirements.txt
pillow>=9.5.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyQt6>=6.5.0
exifread>=3.0.0
tqdm>=4.65.0
psutil>=5.9.0
numpy>=1.22.0
opencv-python>=4.5.0
# RAW file processing
rawpy>=0.17.0
# PyTorch with MPS (Metal Performance Shaders) support
torch>=2.0.0
# CoreML for Metal acceleration
coremltools>=6.0
# macOS-specific libraries for Apple RAW support
pyobjc>=9.0.1  # For CoreImage API access
pyobjc-framework-Quartz>=9.0.1  # For CoreImage processing
# Database support
# sqlite3 is part of the Python standard library
# Visualization enhancements
scipy>=1.10.0
cartopy>=0.21.0  # For map visualizations
# functools.lru_cache is part of the Python standard library
# Apple-optimized TensorFlow (optional)
# tensorflow-macos>=2.9.0
# tensorflow-metal>=0.7.0
